move scrapers into a folder
